[MM-11280] Move logic around to simplify components
