+++ tests/components/activity_log_modal/activity_log.test.jsx
@@ -3,23 +3,23 @@
 
 import React from 'react';
 import {shallow} from 'enzyme';
+import {FormattedMessage} from 'react-intl';
 
 import {General} from 'mattermost-redux/constants';
 
+import {localizeMessage} from 'utils/utils.jsx';
 import ActivityLog from 'components/activity_log_modal/components/activity_log.jsx';
 
 describe('components/activity_log_modal/ActivityLog', () => {
     const baseProps = {
         index: 0,
         locale: General.DEFAULT_LOCALE,
-        currentSession: {props: {os: 'Linux', platform: 'Linux', browser: 'Desktop App'}, id: 'sessionId'},
-        lastAccessTime: new Date(Date.UTC(2018, 8, 20, 12, 15)),
-        firstAccessTime: new Date(Date.UTC(2018, 6, 20, 12, 15)),
-        devicePlatform: 'Linux',
-        devicePicture: 'fa fa-linux',
-        deviceTitle: 'Linux Icon',
-        moreInfo: false,
-        handleMoreInfo: jest.fn(),
+        currentSession: {
+            props: {os: 'Linux', platform: 'Linux', browser: 'Desktop App'},
+            id: 'sessionId',
+            create_at: 1534917291042,
+            last_activity_at: 1534917643890,
+        },
         submitRevoke: jest.fn(),
     };
 
@@ -31,19 +31,54 @@ describe('components/activity_log_modal/ActivityLog', () => {
         expect(wrapper).toMatchSnapshot();
     });
 
-    test('functions are called correctly', () => {
+    test('submitRevoke is called correctly', () => {
         const wrapper = shallow(
             <ActivityLog {...baseProps}/>
         );
-
-        wrapper.instance().handleMoreInfo();
-        expect(baseProps.handleMoreInfo).toBeCalled();
-        expect(baseProps.handleMoreInfo).toHaveBeenCalledTimes(1);
-        expect(baseProps.handleMoreInfo).toBeCalledWith(0);
 
         wrapper.instance().submitRevoke('e');
         expect(baseProps.submitRevoke).toBeCalled();
         expect(baseProps.submitRevoke).toHaveBeenCalledTimes(1);
         expect(baseProps.submitRevoke).toBeCalledWith('sessionId', 'e');
+    });
+
+    test('handleMoreInfo updates state correctly', () => {
+        const wrapper = shallow(
+            <ActivityLog {...baseProps}/>
+        );
+
+        wrapper.instance().handleMoreInfo();
+        expect(wrapper.state()).toEqual({moreInfo: true});
+    });
+
+    test('should match when isMobileSession is called', () => {
+        const wrapper = shallow(
+            <ActivityLog {...baseProps}/>
+        );
+
+        const isMobileSession = wrapper.instance().isMobileSession;
+        expect(isMobileSession({device_id: 'apple'})).toEqual(true);
+        expect(isMobileSession({device_id: 'android'})).toEqual(true);
+        expect(isMobileSession({device_id: 'none'})).toEqual(false);
+    });
+
+    test('should match when mobileSessionInfo is called', () => {
+        const wrapper = shallow(
+            <ActivityLog {...baseProps}/>
+        );
+
+        const mobileSessionInfo = wrapper.instance().mobileSessionInfo;
+
+        const apple = {devicePicture: 'fa fa-apple', deviceTitle: localizeMessage('device_icons.apple', 'Apple Icon'), devicePlatform: <FormattedMessage defaultMessage='iPhone Native Classic App' id='activity_log_modal.iphoneNativeClassicApp' values={{}}/>}; //eslint-disable-line react/jsx-max-props-per-line
+        expect(mobileSessionInfo({device_id: 'apple'})).toEqual(apple);
+
+        const android = {devicePicture: 'fa fa-android', deviceTitle: localizeMessage('device_icons.android', 'Android Icon'), devicePlatform: <FormattedMessage defaultMessage='Android Native Classic App' id='activity_log_modal.androidNativeClassicApp' values={{}}/>}; //eslint-disable-line react/jsx-max-props-per-line
+        expect(mobileSessionInfo({device_id: 'android'})).toEqual(android);
+
+        const appleRN = {devicePicture: 'fa fa-apple', deviceTitle: localizeMessage('device_icons.apple', 'Apple Icon'), devicePlatform: <FormattedMessage defaultMessage='iPhone Native App' id='activity_log_modal.iphoneNativeApp' values={{}}/>}; //eslint-disable-line react/jsx-max-props-per-line
+        expect(mobileSessionInfo({device_id: 'apple_rn'})).toEqual(appleRN);
+
+        const androidRN = {devicePicture: 'fa fa-android', deviceTitle: localizeMessage('device_icons.android', 'Android Icon'), devicePlatform: <FormattedMessage defaultMessage='Android Native App' id='activity_log_modal.androidNativeApp' values={{}}/>}; //eslint-disable-line react/jsx-max-props-per-line
+        expect(mobileSessionInfo({device_id: 'android_rn'})).toEqual(androidRN);
     });
 });
